Refactor: auth로직 분리
