--- src/main/kotlin/team/react/school_chat/domain/auth/repository/FriendRepository.kt
@@ -1,8 +1,0 @@
-package team.react.school_chat.domain.auth.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import team.react.school_chat.domain.auth.domain.Friend
-import team.react.school_chat.domain.auth.domain.FriendId
-
-interface FriendRepository: JpaRepository<Friend, FriendId> {
-}
