+++ src/main/kotlin/team/react/school_chat/domain/member/dto/FriendDTO.kt
@@ -1,8 +1,8 @@
-package team.react.school_chat.domain.auth.dto
+package team.react.school_chat.domain.member.dto
 
-import team.react.school_chat.domain.auth.domain.Friend
-import team.react.school_chat.domain.auth.domain.FriendId
-import team.react.school_chat.domain.auth.domain.Member
+import team.react.school_chat.domain.member.domain.Friend
+import team.react.school_chat.domain.member.domain.FriendId
+import team.react.school_chat.domain.member.domain.Member
 
 data class FriendDTO(
     val friendEmail: String,
