+++ src/main/kotlin/team/react/school_chat/domain/member/dto/MemberDTO.kt
@@ -1,6 +1,6 @@
-package team.react.school_chat.domain.auth.dto
+package team.react.school_chat.domain.member.dto
 
-import team.react.school_chat.domain.auth.domain.Member
+import team.react.school_chat.domain.member.domain.Member
 
 data class MemberDTO(
     val email: String,
