+++ src/main/kotlin/team/react/school_chat/domain/member/dto/MemberUpdateRequest.kt
@@ -1,0 +1,6 @@
+package team.react.school_chat.domain.member.dto
+
+data class MemberUpdateRequest(
+    val name: String,
+    val picture: String
+)
