+++ src/main/kotlin/team/react/school_chat/domain/member/repository/MemberRepository.kt
@@ -1,7 +1,7 @@
-package team.react.school_chat.domain.auth.repository
+package team.react.school_chat.domain.member.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
-import team.react.school_chat.domain.auth.domain.Member
+import team.react.school_chat.domain.member.domain.Member
 
 interface MemberRepository: JpaRepository<Member, String> {
 }
