+++ src/main/kotlin/team/react/school_chat/global/auth/service/CustomOAuth2UserService.kt
@@ -1,4 +1,4 @@
-package team.react.school_chat.domain.auth.service
+package team.react.school_chat.global.auth.service
 
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.core.authority.SimpleGrantedAuthority
@@ -8,10 +8,10 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserService
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User
 import org.springframework.security.oauth2.core.user.OAuth2User
 import org.springframework.stereotype.Service
-import team.react.school_chat.domain.auth.domain.Member
-import team.react.school_chat.domain.auth.repository.MemberRepository
+import team.react.school_chat.domain.member.domain.Member
+import team.react.school_chat.domain.member.repository.MemberRepository
 import team.react.school_chat.global.utils.JwtProvider
-import team.react.school_chat.global.utils.OAuthAttributes
+import team.react.school_chat.global.auth.OAuthAttributes
 import java.util.*
 
 @Service
